+++ Library/ISCore/ISQueryModelHelper.cpp
@@ -1,7 +1,7 @@
 #include "ISQueryModelHelper.h"
 #include "ISConstants.h"
 //-----------------------------------------------------------------------------
-QString ISQueryModelHelper::GetForeignViewNameField(const QString &MetaTableForeignAlias, PMetaForeign *MetaForeign, int Iterator)
+QString ISQueryModelHelper::GetForeignViewNameField(const QString &MetaTableForeignAlias, PMetaForeign *MetaForeign, size_t Iterator)
 {
 	QStringList StringList = MetaForeign->ForeignViewNameField.split(';');
 	if (StringList.count() > 1)
@@ -18,7 +18,7 @@ QString ISQueryModelHelper::GetForeignViewNameField(const QString &MetaTableFore
 	return ISQueryModelHelper::GetAliasForLeftJoinTable(MetaTableForeignAlias, Iterator) + SYMBOL_POINT + MetaTableForeignAlias + '_' + MetaForeign->ForeignViewNameField.toLower();
 }
 //-----------------------------------------------------------------------------
-QString ISQueryModelHelper::GetAliasForLeftJoinTable(const QString &TableAlias, int Iterator)
+QString ISQueryModelHelper::GetAliasForLeftJoinTable(const QString &TableAlias, size_t Iterator)
 {
 	return TableAlias + QString::number(Iterator);
 }
