+++ Library/ISCore/ISQueryModelHelper.h
@@ -6,7 +6,7 @@
 class ISQueryModelHelper
 {
 public:
-	static QString GetForeignViewNameField(const QString &MetaTableForeignAlias, PMetaForeign *MetaForeign, int Iterator); //Получить текст запроса для отображаемого поля эскортной таблицы (Атрибут в схеме ForeignViewNameField)
-	static QString GetAliasForLeftJoinTable(const QString &TableAlias, int Iterator); //Получить алиас для таблицы в связке LEFT JOIN
+	static QString GetForeignViewNameField(const QString &MetaTableForeignAlias, PMetaForeign *MetaForeign, size_t Iterator); //Получить текст запроса для отображаемого поля эскортной таблицы (Атрибут в схеме ForeignViewNameField)
+	static QString GetAliasForLeftJoinTable(const QString &TableAlias, size_t Iterator); //Получить алиас для таблицы в связке LEFT JOIN
 };
 //-----------------------------------------------------------------------------
