+++ Projects/Configurator/CGConfiguratorCreate.cpp
@@ -24,7 +24,7 @@ CGConfiguratorCreate::~CGConfiguratorCreate()
 bool CGConfiguratorCreate::resources()
 {
 	bool Result = true, Exist = true;
-	for (int i = 0, CountResources = ISMetaData::Instance().GetResources().size(); i < CountResources; ++i)
+	for (size_t i = 0, CountResources = ISMetaData::Instance().GetResources().size(); i < CountResources; ++i)
 	{
 		PMetaResource *MetaResource = ISMetaData::Instance().GetResources().at(i);
 		Progress("Resources for " + MetaResource->TableName, i, CountResources);
